+++ workers/bundler.js
@@ -1,0 +1,1 @@
+!function(){"use strict";let e,t;self.window=self;let s=new Promise(e=>t=e);self.addEventListener("message",async o=>{switch(o.data.type){case"init":e=o.data.svelteUrl,importScripts(`${e}/compiler.js`,o.data.rollupUrl),t();break;case"bundle":if(0===o.data.components.length)return;await s;const i=await async function({id:t,components:s}){console.log(`running Svelte compiler version %c${svelte.VERSION}`,"font-weight: bold");const o={};s.forEach(e=>{const t=`./${e.name}.${e.type}`;o[t]=e});const i=new Map;let c;try{if((c=await async function(t,s,n){let o;const i=[],c={};try{o=await rollup.rollup({input:"./App.svelte",external:e=>"."!==e[0]&&(!a(e)&&(!e.startsWith("http://")&&!e.startsWith("https://"))),plugins:[{resolveId(t,s){if("svelte"===t)return`${e}/index.mjs`;if(t.startsWith("svelte/"))return`${e}/${t.slice(7)}.mjs`;if(s&&(s.startsWith("https://")||s.startsWith("http://")))return new URL(`${t}.mjs`,s).href;if(t.endsWith(".html")&&(t=t.replace(/\.html$/,".svelte")),t in n)return t;throw new Error(`Could not resolve "${t}" from "${s}"`)},load:e=>e.startsWith("https://")||e.startsWith("http://")?l(e):e in n?n[e].source:void 0,transform(e,n){if(!/\.svelte$/.test(n))return null;const a=n.replace(/^\.\//,"").replace(/\.svelte$/,""),o=s[n]&&s[n].code===e?s[n].result:svelte.compile(e,Object.assign({generate:t,format:"esm",name:a,filename:a+".svelte"},r));return c[n]={code:e,result:o},(o.warnings||o.stats.warnings).forEach(e=>{i.push({message:e.message,filename:e.filename,start:e.start,end:e.end})}),o.js}}],inlineDynamicImports:!0,onwarn(e){i.push({message:e.message})}})}catch(e){return{error:e,bundle:null,cache:c,warnings:i}}return{bundle:o,cache:c,error:null,warnings:i}}("dom",n.dom,o)).error)throw c.error;n.dom=c.cache;let s=1;const m=(await c.bundle.generate({format:"iife",name:"SvelteComponent",globals:e=>{const t=`import_${s++}`;return i.set(e,t),t},exports:"named",sourcemap:!0})).output[0],u=null;return{id:t,imports:m.imports,import_map:i,dom:m,ssr:u,warnings:c.warnings,error:null}}catch(e){const s=e;return delete s.toString,{id:t,imports:[],import_map:i,dom:null,ssr:null,warnings:c.warnings,error:Object.assign({},s,{message:s.message,stack:s.stack})}}}(o.data);i&&postMessage(i)}});const r={dev:!0};let n={dom:{},ssr:{}};const a=e=>"svelte"===e||e.startsWith("svelte/"),o=new Map;function l(e){return o.has(e)||o.set(e,fetch(e).then(e=>e.text()).catch(t=>{console.error(t),o.delete(e)})),o.get(e)}}();
